+++ ui/shared/AppError/AppError.tsx
@@ -3,16 +3,20 @@ import React from 'react';
 
 import { route } from 'nextjs-routes';
 
+import config from 'configs/app';
 import getErrorCause from 'lib/errors/getErrorCause';
 import getErrorCauseStatusCode from 'lib/errors/getErrorCauseStatusCode';
 import getErrorObjStatusCode from 'lib/errors/getErrorObjStatusCode';
 import getResourceErrorPayload from 'lib/errors/getResourceErrorPayload';
+import AdBannerContent from 'ui/shared/ad/AdBannerContent';
 
 import AppErrorIcon from './AppErrorIcon';
 import AppErrorTitle from './AppErrorTitle';
 import AppErrorBlockConsensus from './custom/AppErrorBlockConsensus';
 import AppErrorTooManyRequests from './custom/AppErrorTooManyRequests';
 import AppErrorTxNotFound from './custom/AppErrorTxNotFound';
+
+const adBannerConfig = config.features.adsBanner;
 
 interface Props {
   className?: string;
@@ -73,6 +77,8 @@ const AppError = ({ error, className }: Props) => {
       default: {
         const { title, text } = ERROR_TEXTS[String(statusCode)] ?? ERROR_TEXTS[500];
 
+        const adBannerProvider = adBannerConfig.isEnabled ? adBannerConfig.provider : null;
+
         return (
           <>
             <AppErrorIcon statusCode={ statusCode }/>
@@ -87,6 +93,7 @@ const AppError = ({ error, className }: Props) => {
             >
                 Back to home
             </Button>
+            { statusCode === 404 && adBannerProvider && <AdBannerContent mt={ 12 } provider={ adBannerProvider }/> }
           </>
         );
       }
