+++ ui/shared/ad/AdBanner.tsx
@@ -1,15 +1,11 @@
-import { chakra, Skeleton } from '@chakra-ui/react';
+import { chakra } from '@chakra-ui/react';
 import React from 'react';
 
 import config from 'configs/app';
 import { useAppContext } from 'lib/contexts/app';
 import * as cookies from 'lib/cookies';
 
-import AdbutlerBanner from './AdbutlerBanner';
-import CoinzillaBanner from './CoinzillaBanner';
-import GetitBanner from './GetitBanner';
-import HypeBanner from './HypeBanner';
-import SliseBanner from './SliseBanner';
+import AdBannerContent from './AdBannerContent';
 
 const feature = config.features.adsBanner;
 
@@ -22,31 +18,12 @@ const AdBanner = ({ className, isLoading }: { className?: string; isLoading?: bo
     return null;
   }
 
-  const content = (() => {
-    switch (provider) {
-      case 'adbutler':
-        return <AdbutlerBanner/>;
-      case 'coinzilla':
-        return <CoinzillaBanner/>;
-      case 'getit':
-        return <GetitBanner/>;
-      case 'hype':
-        return <HypeBanner/>;
-      case 'slise':
-        return <SliseBanner/>;
-    }
-  })();
-
   return (
-    <Skeleton
+    <AdBannerContent
       className={ className }
-      isLoaded={ !isLoading }
-      borderRadius="none"
-      maxW={ ('adButler' in feature && feature.adButler) ? feature.adButler.config.desktop.width : '728px' }
-      w="100%"
-    >
-      { content }
-    </Skeleton>
+      isLoading={ isLoading }
+      provider={ provider }
+    />
   );
 };
